Added correct version of Tau Veto Filter

+++ ElectroWeakAnalysis/EWKTau/python/tauVetoFilter_cff.py
@@ -5,7 +5,7 @@ PfTauTag = cms.InputTag(""),
 CaloTauTag = cms.InputTag(""),
 PatTauTag = cms.InputTag(""),
 MinEtCut = cms.double(20.),
-MaxNCut = cms.double(1.),
+MaxNCut = cms.int32(1),
 Type = cms.string("Pf")
 )
